wtsa-github/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
playwright==1.44.0
requests==2.32.3
httpx==0.27.0
beautifulsoup4==4.12.3
anthropic==0.28.0
notion-client==2.2.1
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
lxml==5.2.2
